collection-service/requirements.txt
Flask>=2.0
flask-cors>=4.0
pymongo>=4.0
python-dotenv>=1.0
requests>=2.28
